Правки

+++ pages/profile/index.vue
@@ -53,28 +53,60 @@
     			</div>
 
     			<div class="profile-links">
+    				<div class="profile-links__item">
+    					<nuxt-link to="#">
+	    					<p class="title">
+	    						Обратная связь
+	    					</p>    					
+	    				</nuxt-link>	    				
+    				</div>
+    				<div class="profile-links__item">    					
+	    				<nuxt-link :to="{name: 'Auth-ResetPassword', query: { title : 'Изменение пароля'}}">
+	    					<p class="title">
+	    						Сменить пароль
+	    					</p>    					
+	    				</nuxt-link>	    				
+    				</div>
+    				<div class="profile-links__item">    					
+	    				<nuxt-link to="#">
+	    					<p class="title">
+	    						Обратная связь
+	    					</p>    					
+	    				</nuxt-link>
+    				</div>
     				<!-- <nuxt-link class="button button--green" to="#">
     					Обратная связь
-    				</nuxt-link> -->
-    				<nuxt-link class="button button--bordered red" :to="{name: 'Auth-ResetPassword', query: { title : 'Изменение пароля'}}">
-    					Изменить пароль
-    				</nuxt-link>
+    				</nuxt-link> -->    				
 
+    				<!-- <nuxt-link class="button button--bordered red" :to="{name: 'Auth-ResetPassword', query: { title : 'Изменение пароля'}}">
+    					Изменить пароль
+    				</nuxt-link> -->
 
     				<!-- <nuxt-link class="button button--transparent green" to="#">
     					Қазақша нұсқасы
     				</nuxt-link> -->
     			</div>
+    			<div class="profile-logout">
+    				<button class="button button--bordered red" @click="showModal('modal-main')">
+    					Выйти из аккаунта
+    				</button>
+    			</div>
     		</div>
 		</div>
+
+		<modal-main :title="title" :text="text" :img="img" :number="number" :btnText="btnText"></modal-main>
 	</main>
 </template>
 <script>
+	import ModalMain from '~/components/layouts/Modals/modal-main.vue'
 	export default{
 		filters:{
 			formatNumber(value) {
 				return value.replace(/(\d{1})(\d{3})(\d{3})(\d{2})(\d{2})/, "$1 $2 $3 $4 $5");
 			}
+		},
+		components: {
+			ModalMain,	
 		},
 		data(){
 			return{
@@ -86,15 +118,28 @@
 				tradeagent:{
 					name: localStorage.getItem('tradeagent') ? JSON.parse(localStorage.getItem('tradeagent')).employee_name : '',
 					tel: localStorage.getItem('tradeagent') ? JSON.parse(localStorage.getItem('tradeagent')).phone : '',
-				}
+				},
+				title: 'Выход',
+				text: 'Вы действительно хотите выйти из аккаунта?',
+				number:'',
+				img: 'logout',
+				btnText: 'logout',				
 			}
 		},
+		mounted(){
+			this.$nuxt.$on('logOut',this.logOut)
+		},
 		methods:{
+			showModal(modal){
+				console.log('modal', modal)
+				$('#'+modal).modal('show')
+			},
 			async logOut(){
 
 				this.$axios.defaults.headers.common['Authorization'] = 'Bearer '+ localStorage.getItem('authToken');
 				await this.$axios.get('/auth/logout')
 					.then(response =>{
+						$('#modal-main').modal('hide');
 						localStorage.clear();
 						this.$store.commit('resetState');
 						this.$router.push('/')
@@ -158,7 +203,7 @@
 				font-size: 12px;
 				line-height: 14px;
 				color: #969696;
-				margin-bottom: 4px;
+				margin-bottom: 4px;				
 			}
 			.name{
 				font-weight: 500;
@@ -184,7 +229,10 @@
 		// links
 
 		&-links{
-			margin-top: 24px;
+			margin-top: 8px;
+			border-radius: 8px;
+			background: #fff;
+			padding: 0px 16px 0px 16px;
 			.button{
 				&--green{
 					display: block;
@@ -196,6 +244,59 @@
 					display: block;
 				}
 			}
+
+			&__item{
+				position: relative;
+				// margin-bottom: 16px;
+				padding: 18px 0;
+				&:last-child{
+					margin-bottom: 0px;
+					// padding-bottom: 0;
+					&:after{
+						display: none;
+					}
+				}
+				&>a{
+					&:hover{
+						text-decoration: none;
+					}
+				}
+				&:after{
+					content: '';
+				    position: absolute;
+				    width: 100%;
+				    height: 1px;
+				    background: #F0F0F0;
+				    bottom: 0;
+				}
+				p{
+					margin-bottom: 0;
+					&.title{
+						font-weight: 500;
+						font-size: 16px;
+						line-height: 19px;
+						color: #1F1F1F;
+						position: relative;
+						&:after{
+							content: '';
+							position: absolute;
+							top: 0;
+							bottom: 0;
+							right: 0;
+							margin: auto;
+							transform: rotate(180deg);
+							background: url('~assets/img/icons/backlink.svg');
+							width: 12px;
+							height: 20px;
+						}
+					}
+				}
+			}
 		}
+
+		&-logout{
+			margin-top: 24px;
+		}
+		
 	}
 </style>
